# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VeritasMed API",
    "description": "Self-verifying medical literature QA backend",
    "version": "0.8.0"
  },
  "paths": {
...
    "/api/chunk/{chunk_id}": {
      "get": {
        "summary": "Get Chunk",
        "description": "Return the chunk and its neighbouring chunks.",
        "operationId": "get_chunk_api_chunk__chunk_id__get",
        "parameters": [
          {
            "name": "chunk_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Chunk Id"
            }
          },
          {
            "name": "context_window",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 3,
              "minimum": 0,
              "default": 1,
              "title": "Context Window"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChunkContextResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/audit/examples/{response_id}": {
      "get": {
        "tags": [
          "answer audit"
        ],
        "summary": "Replay",
        "operationId": "replay_api_audit_examples__response_id__get",
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          },
          {
            "name": "strategy",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "direct",
                "split",
                "context",
                "quote_v2",
                "atomic_v1",
                "atomic_v2"
              ],
              "type": "string",
              "default": "direct",
              "title": "Strategy"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audit": {
      "post": {
        "tags": [
          "answer audit"
        ],
        "summary": "Live Audit",
        "operationId": "live_audit_api_audit_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtomicV2Request"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AtomicV2Request": {
        "properties": {
          "answer": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Answer"
          },
          "sources": {
            "items": {
              "$ref": "#/components/schemas/AuditSource"
            },
            "type": "array",
            "maxItems": 40,
            "minItems": 1,
            "title": "Sources"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "direct",
              "split",
              "context",
              "quote_v2",
              "atomic_v1",
              "atomic_v2"
            ],
            "title": "Strategy",
            "default": "direct"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "answer",
          "sources"
        ],
        "title": "AtomicV2Request"
      },
      "AuditSource": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Id"
          },
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "text": {
            "type": "string",
            "maxLength": 50000,
            "minLength": 1,
            "title": "Text"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "title",
          "text"
        ],
        "title": "AuditSource"
      },
      "ChunkContextResponse": {
        "properties": {
          "chunk": {
            "$ref": "#/components/schemas/ChunkSlim"
          },
          "prev_chunk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChunkSlim"
              },
              {
                "type": "null"
              }
            ]
          },
          "next_chunk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ChunkSlim"
              },
              {
                "type": "null"
              }
            ]
          },
          "document": {
            "additionalProperties": true,
            "type": "object",
            "title": "Document"
          }
        },
        "type": "object",
        "required": [
          "chunk",
          "document"
        ],
        "title": "ChunkContextResponse"
      },
...
      "ChunkSlim": {
        "properties": {
          "chunk_id": {
            "type": "string",
            "title": "Chunk Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "chunk_id",
          "text"
        ],
        "title": "ChunkSlim"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}